requirements.txt
Django==4.2
djangorestframework==3.16.1
psycopg2-binary==2.9.7
python-dotenv
